# optimize SupersetOf func performance