**Please provide a description of this PR:**

1. Optimize `SupersetOf` func performance
2. Add an alias func `ContainsAll` of `SupersetOf`
3. Add more test cases

benchmark:
```shell
$ go test -bench="BenchmarkSupersetOf" -test.benchmem -count=10 . >> old.txt # old code
$ go test -bench="BenchmarkSupersetOf" -test.benchmem -count=10 . >> new.txt # new code
$ benchstat old.txt new.txt
name                      old time/op    new time/op    delta
SupersetOf/SupersetOf-12    1.28µs ± 4%    1.17µs ± 6%    -8.44%  (p=0.000 n=9+10)

name                      old alloc/op   new alloc/op   delta
SupersetOf/SupersetOf-12     48.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
SupersetOf/SupersetOf-12      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```